+++ src/amuse/io/store_v2.py
@@ -1,3 +1,10 @@
+try:
+    import h5py
+except ImportError as ex:
+    import warnings
+    warnings.warn("could not load h5py, hdf5 files not supported", ImportWarning)
+    h5py = None
+    
 import h5py
 import numpy
 import pickle
@@ -570,6 +577,9 @@ class StoreHDF(object):
     DATA_GROUP_NAME = 'data'
     
     def __init__(self, filename, append_to_file=True, open_for_writing = True, copy_history = False, return_working_copy = False):
+        if h5py is None:
+            raise AmuseException("h5py module not available, cannot use hdf5 files")
+            
         if not append_to_file and open_for_writing and os.path.exists(filename):
             os.remove(filename)
